% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_manifold_content.R
\name{clean_manifold_content}
\alias{clean_manifold_content}
\title{Convert API response to a data frame}
\usage{
clean_manifold_content(input)
}
\arguments{
\item{input}{Output from manifold_api() or other convenience function. Output from queries regarding groups, comments, or bets are not currently compatible with this function.}
}
\value{
A tibble
}
\description{
An optional convenience function to format certain results from the Manifold API
}
\examples{
clean_manifold_content(manifold_api("/v0/user/ManifoldMarkets", request_type = "GET"))
}
