% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_sf.r
\name{Drawing}
\alias{Drawing}
\title{Read a Drawing from a Manifold project.}
\usage{
Drawing(mapfile, dwgname, ...)
}
\arguments{
\item{mapfile}{Manifold project file}

\item{dwgname}{Drawing name to read}

\item{...}{passed on to \code{\link{mfd_read_db}}}
}
\value{
Drawing returns a `sf` object
}
\description{
Read a Drawing from a Manifold project.
}
\examples{
mapfile <- system.file("extdata", "AreaDrawing.map", package = "manifoldr")
#dwg  <- Drawing(mapfile, "Drawing")
## only lines
#dwg_sub <- Drawing(mapfile, WHERE = "WHERE [Type (I)] = 2")
#dwg_sub
}

