% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manifold_api_key.R
\name{get_manifold_api_key}
\alias{get_manifold_api_key}
\title{Storing and referencing your Manifold Markets API Key}
\usage{
get_manifold_api_key()
}
\value{
API key as string
}
\description{
Storing and referencing your Manifold Markets API Key
}
\examples{
get_manifold_api_key()
}
