% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market.R
\name{get_market}
\alias{get_market}
\title{get_market()}
\usage{
get_market(market_id_or_slug)
}
\arguments{
\item{market_id_or_slug}{A slug or ID associated with a single market}
}
\value{
Manifold API object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieve a specific market
}
\examples{
get_market("pOIULRJ4zbYyk6dYEmzl")
get_market("will-democrats-lose-their-majoritie")
}
