% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markets.R
\name{get_markets}
\alias{get_markets}
\title{get_markets()}
\usage{
get_markets(n_markets = 100)
}
\arguments{
\item{n_markets}{Number of markets to retrieve}
}
\value{
A JSON object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieve recent markets from most to least recent.
}
\examples{
get_markets(2500)
}
