% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_balance.R
\name{get_my_balance}
\alias{get_my_balance}
\title{get_my_balance()}
\usage{
get_my_balance(key = get_manifold_api_key())
}
\arguments{
\item{key}{A Manifold Markets API key for authentication}
}
\value{
Double
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieve the authenticated user's mana balance.
}
