% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{get_users()}
\usage{
get_users(n_users = 100)
}
\arguments{
\item{n_users}{Number of users to retrieve}
}
\value{
Manifold API object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieve recent users from most to least recent.
}
\examples{
get_users(5)
}
