% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.r
\name{Image}
\alias{Image}
\title{Read an Image from a Manifold project file.}
\usage{
Image(mapfile, rastername)
}
\arguments{
\item{mapfile}{Manifold project file}

\item{rastername}{Image name to read}
}
\value{
RasterBrick
}
\description{
For now we are assuming we just get an RGB 3-layer object.
}
\examples{
fmap <- "V20160202016022.L3m_R3QL_NPP_CHL_chlor_a_9km.map"
mapfile <- system.file("extdata", fmap, package= "manifoldr")
#im <- Image(mapfile, "V20160202016022.L3m_R3QL_NPP_CHL_chlor_a_9km")
}

