% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_api.R
\name{manifold_api}
\alias{manifold_api}
\title{Manifold API}
\usage{
manifold_api(path, key = NULL)
}
\arguments{
\item{path}{One of the endpoints listed at https://docs.manifold.markets/api#get-v0users}

\item{key}{Optional: provide API key}
}
\value{
A JSON object
}
\description{
Interact with the Manifold Markets API
}
\examples{
manifold_api("/v0/user/ManifoldMarkets")
manifold_api("/v0/me", key = manifold_api_key())
}
