% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.r
\docType{class}
\name{ManifoldODBCConnection-class}
\alias{ManifoldODBCConnection-class}
\title{Class ODBCConnection.}
\description{
\code{ODBCConnection} objects are usually created by \code{\link[DBI]{dbConnect}}
}
\keyword{internal}

