% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.r
\docType{class}
\name{ManifoldODBCDriver-class}
\alias{ManifoldODBC}
\alias{ManifoldODBCDriver-class}
\title{ManifoldODBCDriver and methods.}
\usage{
ManifoldODBC()
}
\description{
ManifoldODBCDriver and methods.

Generate an object of ManifoldODBCDriver class
}
\keyword{internal}

