% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldr-package.R
\docType{package}
\name{manifoldr-package}
\alias{manifoldr}
\alias{manifoldr-package}
\title{manifoldr: Manifold Markets API Wrapper}
\description{
The R package `manifoldr` is an API wrapper for the prediction market platform "Manifold Markets." It allows R users to more easily access the Manifold Markets API to create markets, place bets on existing markets, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcblsn/manifoldr}
  \item Report bugs at \url{https://github.com/jcblsn/manifoldr/issues}
}

}
\author{
\strong{Maintainer}: Jacob Eliason \email{jacobeliason@gmail.com}

}
\keyword{internal}
