% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_sf.r
\name{mfd_read_db}
\alias{mfd_read_db}
\title{Read simple features from the Manifold driver}
\usage{
mfd_read_db(con = NULL, table, query = "", geom_column = NULL,
  WHERE = "", ..., quiet = TRUE)
}
\arguments{
\item{con}{connection object}

\item{table}{table to read, can be a Drawing name or its child table}

\item{query}{query to run, optional}

\item{geom_column}{this is the name of the column as returned (not the name of the column natively)}

\item{WHERE}{optional WHERE clause, as in "WHERE ID = 1 AND size < 10"}

\item{...}{other arguments passed along (none currently)}

\item{quiet}{keep quiet by default}
}
\description{
Read a drawing from a Manifold connection.
}
\details{
By default the entire drawing is read with all non-intrinsic columns. The intrinsic column "Geom (I)" is
cast to WKB within Manifold and then interpreted using \code{\link[sf]{st_as_sfc}}.
}

