% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odbc.R
\name{odbcConnectManifold}
\alias{odbcConnectManifold}
\title{ODBC connection to Manifold map files.}
\usage{
odbcConnectManifold(mapfile)
}
\arguments{
\item{mapfile}{}
}
\value{
RODBC object
}
\description{
Create an ODBC connection for Manifold GIS.
}
\details{
See \code{\link[RODBC]{odbcDriverConnect}}
}

