% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{src_manifold}
\alias{src_desc.src_manifold}
\alias{src_manifold}
\title{Manifold source}
\usage{
src_manifold(dbname = NULL, host = NULL, port = NULL, user = NULL,
  password = NULL, ...)

\method{src_desc}{src_manifold}(con)
}
\arguments{
\item{dbname}{}

\item{host}{}

\item{port}{}

\item{user}{}

\item{password}{}

\item{...}{}
}
\description{
Manifold source
}
\examples{
\dontrun{
mapfile <- system.file("extdata", "AreaDrawing.map", package = "manifoldr")
mfd <- src_manifold(mapfile)
tbl(mfd, "GEOMETRY_COLUMNS")
}
}

