% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_bet.R
\name{submit_bet}
\alias{submit_bet}
\title{submit_bet}
\usage{
submit_bet(
  mm_key = get_manifold_api_key(),
  mm_amt,
  mm_market_id = NULL,
  mm_market_slug = NULL,
  mm_outcome = c("YES", "NO"),
  mm_limit = NULL
)
}
\arguments{
\item{mm_key}{A Manifold Markets API key for authentication}
}
\value{
Manifold API object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Create a new market
}
