% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldr.r
\name{Surface}
\alias{Surface}
\title{Read a Surface from a Manifold project file.}
\usage{
Surface(mapfile, rastername)
}
\arguments{
\item{mapfile}{Manifold project file}

\item{rastername}{Surface name to read}
}
\value{
RasterLayer
}
\description{
Read a Surface from a Manifold project file.
}
\examples{
mapfile2 <- system.file("extdata", "Montara_20m.map", package= "manifoldr")
gg <- Surface(mapfile2, "Montara")
}

