% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{tbl.src_manifold}
\alias{tbl.src_manifold}
\title{Tibble method for Manifold}
\usage{
\method{tbl}{src_manifold}(src, from, ...)
}
\arguments{
\item{src}{}

\item{from}{}

\item{...}{}
}
\description{
Tibble method for Manifold
}

