% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldr.r
\name{wktCRS2proj4}
\alias{wktCRS2proj4}
\title{#' @importFrom RODBC sqlQuery
manifoldDrawingCRS <- function(connection, componentname) {
   qu <- sprintf('SELECT TOP 1 CoordSysToWKT(CCoordSys([Geom (I)])) AS [CRS] FROM [%s]',  componentname)
  RODBC::sqlQuery(connection, qu, stringsAsFactors = FALSE)$CRS
}}
\usage{
wktCRS2proj4(CRS)
}
\arguments{
\item{CRS}{coordinate reference system in WKT form}
}
\description{
#' @importFrom RODBC sqlQuery
manifoldRasterCRS <- function(connection, componentname) {
  ## this should work but does not
  ##qu <- sprintf('SELECT TOP 1 CoordSysToWKT(CoordSys("%s" AS COMPONENT)) AS [CRS] FROM [%s]', componentname, componentname)
 # TOP 1 CCoordSys(NewPoint([Easting (I)], [Northing (I)]))
  qu <- sprintf('OPTIONS COORDSYS("[%s]" AS COMPONENT);SELECT TOP 1 CoordSysToWKT(CCoordSys(NewPoint([Easting (I)], [Northing (I)]))) AS [CRS] FROM [%s];',  componentname, componentname)
  #print(qu)
  RODBC::sqlQuery(connection, qu, stringsAsFactors = FALSE)$CRS
}
}

