% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_manifold_content.R
\name{clean_manifold_content}
\alias{clean_manifold_content}
\title{Convert API response to a data frame}
\usage{
clean_manifold_content(input)
}
\arguments{
\item{input}{The result from manifold_api() or other}
}
\value{
A data frame
}
\description{
An optional helper function to clean up the result from the Manifold API
}
\examples{
clean_manifold_content(manifold_api("/v0/user/ManifoldMarkets"))
}
