% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manifold_api_key.R
\name{get_manifold_api_key}
\alias{get_manifold_api_key}
\title{Storing and referencing your Manifold Markets API Key}
\usage{
get_manifold_api_key()
}
\value{
API key as string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
get_manifold_api_key()
}
