% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_api.R
\name{manifold_api}
\alias{manifold_api}
\title{Manifold API}
\usage{
manifold_api(path, key = NULL, params_list = NULL)
}
\arguments{
\item{path}{One of the endpoints listed in Manifold's [documentation](https://docs.manifold.markets/api)}

\item{key}{Optional: provide API key}

\item{params_list}{Optional: provide a list of parameters names and values (see doc for syntax)}
}
\value{
A JSON object
}
\description{
Interact with the Manifold Markets API
}
\examples{
manifold_api("/v0/user/ManifoldMarkets")
manifold_api("/v0/me", key = manifold_api_key())
manifold_api(path = "/v0/markets", params_list = list("limit" = "10", "before" = "EvIhzcJXwhL0HavaszD7"))
}
