% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_api.R
\name{manifold_api}
\alias{manifold_api}
\title{Manifold API}
\usage{
manifold_api(
  endpoint,
  request_type = c("GET", "POST"),
  key = NULL,
  params_list = NULL
)
}
\arguments{
\item{endpoint}{One of the endpoints listed in Manifold's \href{https://docs.manifold.markets/api}{documentation}}

\item{request_type}{Must be one of "GET" or "POST" per endpoint documentation}

\item{key}{API key (see documentation for which endpoints require authentication)}

\item{params_list}{A list of parameter names and values (see documentation for syntax)}
}
\value{
A JSON object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Interact with the Manifold Markets API
}
\examples{
manifold_api(
  endpoint = "/v0/user/ManifoldMarkets",
  request_type = "GET"
)
manifold_api(
  endpoint = "/v0/me",
  key = get_manifold_api_key(),
  request_type = "GET"
)
manifold_api(
  endpoint = "/v0/markets",
  request_type = "GET",
  params_list = list(limit = 10, before = "EvIhzcJXwhL0HavaszD7")
)
}
