% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_api.R
\name{manifold_api}
\alias{manifold_api}
\title{Manifold API}
\usage{
manifold_api(
  endpoint,
  request_type = c("GET", "POST"),
  key = NULL,
  params_list = NULL
)
}
\arguments{
\item{endpoint}{One of the endpoints listed in Manifold's [documentation](https://docs.manifold.markets/api)}

\item{request_type}{Must be one of "GET" or "POST" per endpoint documentation}

\item{key}{API key (see documentation for which endpoints require authentication)}

\item{params_list}{A list of parameter names and values (see documentation for syntax)}
}
\value{
A JSON object
}
\description{
Interact with the Manifold Markets API
}
\examples{
manifold_api("/v0/user/ManifoldMarkets")
manifold_api("/v0/me", key = get_manifold_api_key())
manifold_api(endpoint = "/v0/markets", params_list = list(limit = 10, before = "EvIhzcJXwhL0HavaszD7"))
}
