% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldr.r
\name{readmfd}
\alias{readmfd}
\title{manifoldr-internal}
\usage{
readmfd(dsn, table, WHERE = NULL, spatial = FALSE, topol = c("area",
  "line", "point"), crs = NULL)
}
\arguments{
\item{dsn}{data source}

\item{table}{table name}

\item{WHERE}{where clause/s}

\item{spatial}{true or false}

\item{topol}{area, line, or point}

\item{crs}{NULL by default, can pass in NA_character_ to avoid CRS complaints}
}
\description{
manifoldr-internal
}

