% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifoldr.r
\name{odbcConnectManifold}
\alias{odbcConnectManifold}
\title{ODBC connection to Manifold map files.}
\usage{
odbcConnectManifold(mapfile, unicode = TRUE, ansi = TRUE, opengis = TRUE)
}
\arguments{
\item{mapfile}{Manifold project *.map file}
}
\value{
RODBC object
}
\description{
Create an ODBC connection for Manifold GIS.
}
\details{
See \code{\link[RODBC]{odbcDriverConnect}}
}
\examples{
\dontrun{
f <- system.file("extdata", "AreaDrawing.map", package = "manifoldr")
con <- odbcConnectManifold(f)
tab <- RODBC::sqlQuery(con, "SELECT * FROM [Drawing]")
## drop [Geom (I)] and give a summary
summary(subset(tab, select = -`Geom (I)`))

## issue a spatial query
qtx <- "SELECT [ID], [Name], [Length (I)] AS [Perim], 
     BranchCount([ID]) AS [nbranches] FROM [Drawing Table]"
sq <- RODBC::sqlQuery(con, qtx)
sq

}
}

